Moved FormSecondPage to its own folder, added its CSS file, updated README and refactored imports for newer structure
